State Codec (#77)
* Update test vectors and modify them to match the corresponding types.
* Merge `AvailabilitySpecifier` and `CAvailabilitySpecifier`, and correct the key for reading JSON.
* Change the `ImportSegment`'s `uint32` to `uint16`, and also fixed the type in the `node`.
* State encoding completed.
* Fixed the issue in MarshalJSON() that caused an error in TestNodeSafrole.
* Fixed the issue where `MarshalJSON()` encoded the hash as a hex string, which caused an error in `TestNodeSafrole`.
) `MarshalJSON()` will only print the hash as a hex string, without encoding the content as a hex string as well .However, now the hex content is printed at the front of the `Marshal` structure*."

---------

+++ node/node.go
@@ -995,7 +995,7 @@ func (n *Node) processAvailabilityJustification(aj *types.AvailabilityJustificat
 	return nil
 }
 
-func (n *Node) getImportSegment(treeRoot common.Hash, segmentIndex uint32) ([]byte, error) {
+func (n *Node) getImportSegment(treeRoot common.Hash, segmentIndex uint16) ([]byte, error) {
 	// TODO: do you need segmentRoot or segmentsRoot here?
 	fmt.Printf("treeRoot: %v, segmentIndex: %v \n", treeRoot, segmentIndex)
 	segmentData, err := n.FetchAndReconstructSpecificSegmentData(treeRoot)
